use resize observer to detect changes in the webview layout

+++ BeyondWordsPlayer/BeyondWordsPlayer/BridgeMessage.swift
@@ -9,6 +9,6 @@ struct BridgeMessage: Decodable {
     var type: String
     var event: PlayerEvent? = nil
     var settings: PlayerSettings? = nil
-    var width: Int? = nil
-    var height: Int? = nil
+    var width: Float? = nil
+    var height: Float? = nil
 }
